added FilterGraph for AudioPluginHost
When you scan for the plugin in APH, be sure to remove any plugins that were built with Thread Sanitizer enabled.  Scanning with APH as the plugin executable will fail with plugins that have Thread Sanitizer turned on, if APH and this project aren't built with Thread Sanitizer turned on.